build: local development settings
previously when editing settings for local development, files under version
control needed to be touched. that then required ongoing care-taking and
version control management which was cumbersome.

change is to initialize config as well from "setting.local.js" if it exists
which has the local development settings.

the example file "setting.local.dist.js" is to be copied over to
"setting.local.js" in the same directory to easily set-up development.

now there is no need to edit "setting.js" to configure the seatable-server,
or "locale/index.js" to change the language for development any longer.

NOTE: window.dtablePluginConfig renamed to window.dtable