build: dtable-ui window.app.expandRow() dev shim
the window.app.expandRow() function is in use to open the "Row Details"
editor in the calendar and other plugins, see [ref-sdk].

in a development setup this part of the dtable-sdk dtable-ui API is not
available.

adding a shim in the development setup allows to track app.expandRow()
initiations via the browser console.

ref-sdk: https://docs.seatable.io/published/dtable-sdk/dtable-ui.md
forum: https://forum.seatable.io/t/how-does-row-details-work-in-the-calendar-plugin/321