fix all-day for all-day event
previously event.allDay was unset, always, even for all-day events.

not setting the the allDay property of an event to true when it is in fact
an all-day-event makes it a zero-duration event in the browsers timezone.

fix is to set the allDay property to true if the constructing start date
is a date without minute precision.

[RBC-Doc]: http://jquense.github.io/react-big-calendar/examples/index.html#prop-events