fix event end <= start / minimum length
previously the start and end of an event was accepted if truthy for the
date columns in the table event row.

this results in display failures and erroneous behavior when the events'
end is before the start with the vendored react-big-calendar
implementation on multiple calendar views including year and month.

fix is to accept the end only if it is greater than the start for non-all-
day events. if not, the end is the start plus one hour.

this solves the immediate display problem. moving the event will persist
the corrected event end in the event table row.

this also solves an issue with the week and day views in specific when
the row end date was not set as the react-big-calendar displays the event
within the slot immediately before its start.