all-/cross- day event
basically, given the end-date is undefined|null|"", the event is all-day.

additionally an event is displayed in the all-day slot if it is across
multiple days.

previously there was no need to handle all- or cross-days to this extend
for any calendar event as only the original week view was in use and
additionally the year view.

now this is needed as the week and day views are active and they display
event with time(s) and within the all-/cross-day slot on top.

this is handling events with a bogus end-date gracefully:

in case an end-date is before the start-date, the start-date is leading
and the end-date of the event will be the events start-date plus a fixed
amount of time (e.g. one hour). this is useful for events with date and
time (with minute precision).

and an offset of 12 hours is necessary to display the event on the
correct day within all views, incl. the year view.