build: keep backups of plugin-zip builds
previously when building the zip (build-plugin), the zip-file would have been
overwritten if not eradicated by (clean) earlier.

as the only file produced in the plugin-zip build folder is the zip file
itself that would be overwritten in any case, the (clean) script is
technically superfluous for the (build-plugin) npm script when building
the same version.

in this change-set:

- remove clean from build-plugin to preserve folder contents
- establish the output path per git repository (removes mkdir requirement)
- change clean to use git-clean(1) instead of rm -rf
- build zip revisions (e.g. different plugin versions)
- backup of earlier zip builds (one by default) for same plugin version

this makes the (build-plugin) zip build less destructive by keeping
backups in a similar way as cp(1) --backup=numbered.

- numbered is using a numbered suffix like ".~1~"
- "npm run clean" deletes all zips _and_ backups

this is oriented at the said cp(1) behavior (GNU Linux) [Ref-CP], another
reference for backup files is emacs [Ref-Emacs-1], [Ref-Emacs-2].

compared to before the change, output zip files are not overwritten
silently any more and a backup is kept of the file.

changing the version and/or the project name will create additional zip
files as the version is part of the zip file-name and the output directory
is not cleaned any longer.

to get the old behaviour run clean before build.

NOTE: the zip files' filename changes to a suffixed variant with the plugin
      version (calendar.zip -> calendar-0.1.20.zip).

[Ref-CP]: https://linux.die.net/man/1/cp
[Ref-Emacs-1]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Backup-Names.html
[Ref-Emacs-2]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Numbered-Backups.html#Numbered-Backups