use intl for calendar messages
currently messages are translated by intl at place of message use for the
default locale (en) message _text_.

change ensures the actual messages are translated, by giving them a domain
key in the form of `.rbc.messages.<message-key>`.

this makes translations immune to changes of messages texts in the default
locale (en) and allows to fully translate all messages in all locales.

NOTE: due to a flaw in the high-order-component (hoc) for drag-and-drop
      (dnd) which is in use, the merged localizer (which has the messages)
      is not properly propagated. therefore transparent translation of all
      messages according to locale is unfortunately not yet possible to
      do in the intl-decorator.