switch localizer to moment locales (#44)
* switch localizer to moment locales

previously in the calendar, there was a mixture of using the localizer
formats and intl, also multiple times per a single localization.

in aa37d42 most of these places were already touched and could be
reduced, but they could not yet be fully addressed.

here now for all calendar dates incl. documentation for the changes in
translation.

migrates the following intl keys back to the original localization
formats:

- Week_xxx
- xxx_Year
- xxx_Month_xxx_Date_xxx_Week

The localization of dates can now be translated to specific formats
by overriding the original formats when needed. This is much more
flexible. The following two keys are introduced:

- .rbc.localizer.moment.yearHeaderFormat
- .rbc.localizer.moment.dayHeaderFormat

These keys are special in-so-far that if undefined they will fall-back
to the original format specifier (part after the last dot).

it comes with examples for level of changes:

1. date format per react-intl-universal key
2. date format function by culture *
3. replace of a format with another one for all cultures.

and includes a map of the intl-locale to moment localizer culture, here
specifically intl locale "en" is mapped to localizer culture "en-GB".

NOTE: this hides a problem that on the month view the first day of week
      is monday even in the (en) culture it is sunday as in the en-GB
      culture it is monday. as year and week views are correct to the
      start of the week day, only month was displaying wrong. this is
      hidden now.

* use-case is the week view, specifically the format in the header. your
  branch might not this view.

ref: aa37d42364b1693c96dba5fb58fd3184b08ab025

* translations

* 1.use "ddd" to format weekdays in the month view, use "dd" to format weekdays in the year view.
2.keep Sunday as the first day of the week.

Co-authored-by: renjie-run <rj.aiyayao@gmail.com>