locales: rename weekdayShortFormat to yearMonthWeekdayFormat
in #44 a new named format "weekdayShortFormat" was introduced.

it is specific to the year-month year-widget and not part of the calendar
component standard formats (in the now decorated localizer).

renaming to a more specific yearMonthWeekdayFormat so that it is more
visible for what it is for and better aligned with the original localizer
formats for specific-use formats.