locales: add intl for weekday-formats
add intl translations for the new yearMonthWeekdayFormat and the standard
weekdayFormat.

original weekdayFormat is ddd (was before #44) and changed to dd (in #44)
which showed need for a specific weekday-format for the per-month calendars
in the year-view.

this change builds on top of #44 allowing to add translations for these
formats across all intl locales.

it also allows within the intl-decorator to define new named formats with
translations with a standard fall-back format in case it is undefined in
the subject localizer formats (_formats).