all-/cross- day: drag-and-drop
with the changes to the event model and the new views, dragging and
dropping events have more features now, like turning an event into an
all-day event.

row updates are done more specifically to the columns that actually need
changes.

the events collection is updated specifically, too. previously it was
not updated and re-created after row changes did propagate.

when there are full-day-dates due to start date without minute precision
(all day events), and the end date is also w/o minute precision, then
it needs to be noon (12:00) not start of day (00:00) as otherwise the
day would not count as end (the year view seems most picky).

it is only necessary to modify the end-date to make it into the all-day
slot if it is on the same day.